requirements.txt
cd # external requirements
certifi==2019.9.11
chardet==3.0.4
typer>=0.6.1
Flask>=2.1.0
idna==2.8
jinja2>=2.11.3
pyjwt>=2.4.0
pytz==2019.3
six==1.13.0
twilio>=6.33.1
urllib3>=1.26.5
python-dotenv>=0.5.1
openai
chronological
requests
ipython
ipykernel
pandas
apscheduler
-e openai-chatbot/